+++ src/azure-cli/HISTORY.rst
@@ -2,6 +2,19 @@
 
 Release History
 ===============
+**AMS**
+
+* BREAKING CHANGE:
+    Changed job creation from "job start" to "job create".
+
+2.0.74
+++++++
+
+**ACR**
+
+* Added a required `--type` parameter to command `az acr config retention update`
+* Param `-n, --name` changed to `-r, --registry` for `az acr config` command group.
+
 **AKS**
 
 * Add `--load-balancer-sku`, `--load-balancer-managed-outbound-ip-count`, `--load-balancer-outbound-ips` and `--load-balancer-outbound-ip-prefixes` to `az aks create` command, which allows for creating AKS cluster with SLB.
@@ -11,11 +24,7 @@ Release History
 **ARM**
 
 * az group deployment create: Add --handle-extended-json-format to support multiline and comments in json template
-
-**ACR**
-
-* Added a required `--type` parameter to command `az acr config retention update`
-* Param `-n, --name` changed to `-r, --registry` for `az acr config` command group.
+* Update azure-mgmt-resource package to use 4.0.0
 
 **Compute**
 
@@ -43,6 +52,11 @@ Release History
 **Monitor**
 
 * monitor metrics alert create: Fix #9901. Support special character `:` in `--condition` argument.
+
+**Policy**
+
+* Support for Policy new API version 2019-06-01.
+* az policy assignment create: Add `--enforcement-mode` parameter for `az policy assignment create` command.
 
 **Storage**